+++ apps/web/tsconfig.json
@@ -10,7 +10,7 @@
     "sourceMap": true,
     "strict": true
   },
-  "exclude": ["./node_modules/"]
+  "exclude": ["node_modules/**"]
   // Path aliases are handled by https://kit.svelte.dev/docs/configuration#alias
   //
   // If you want to overwrite includes/excludes, make sure to copy over the relevant includes/excludes
